backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
aiomysql
python-dotenv
pydantic
passlib[bcrypt]
python-multipart
passlib
python-jose[cryptography]
PyJWT
requests
beautifulsoup4
fastapi-mail
nltk

beautifulsoup4
trafilatura
pdfminer.six
python-docx
pandas
openpyxl
xlrd

faiss-cpu
google-genai
langchain
langchain_community
langchain_google_genai
langchain-openai
scikit-learn